Release 2.5.1
Requires updating both the 32u4 sketch and the ESP sketch.

Add a serial link to allow issuing commands from the 32u4 back to the ESP8266, this may be utilized more in future releases.

Add the ability to set an open network via serial for devices that have trouble connecting to a secured ESP8266 access point (rare cases).  This makes it easier for the new users who experience this rare issue to get started using their device.